Delete "required" attributes on password and salt properties in the AppUser class. Add a "Created" property in the AppUser class to track when a user was created.
Add an Accounts table and AccountAppUser table because there is a many-to-many relationship between the AppUser and the Account. Add Account table as DbSet in the Data Context class.

Add navigation properties to Users and Accounts tables, referencing [this article in the Microsoft EF Core documentation](https://learn.microsoft.com/en-us/ef/core/modeling/relationships/many-to-many).

Add seed data in UserSeedData.json file to be read in Seed.cs file. This file gets read and deserialized into a C# AppUser object (neither tracked by Git because they have sensitive data, and their only real purpose is to add seed data so there is data that can be used while building out the front-end). This object will contain the users of the app from the seed file, and the users will get a password and hash assigned to them.

Add migrations and create new table with seed data.