Create new Categories controller with endpoint for retrieving categories from the db.
Update Debits controller with method for adding a debit, but will need to tweak this method to add a new description id and category id if the description and/or category are custom (defined by the submission of a new debit with new category and/or description).

Add Descriptions controller with endpoint for retrieving categories from the db.

Update Debit repository and Description repository (and associated interfaces) with new methods.

Update DebitDto with nullable operator for the description name.

Update AutoMapper class with mapping from DebitDto to Debit entity.

Add debit service with API endpoints.

Update debits component and template with methods for retrieving descriptions and categories and adding a new debit.